+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - const/perms: useful path permissions constants 
 - data/filerw : an object for basic read/write operations on a file
 - db/sqlite: an object wrapper around the standard SQLITE3 library with a logger
-- fs/dir : a directory object with a wipe method to remove the tree (while protecting root and system files)
+- fs/dir : a directory object (subfolders, files, wiping directories)
 - fs/groups : get the group name by gid or GID for a group name
 - fs/pathperms: Path permission helpers
 - fs/users: get the user name by uid or UID for a user name